move from registry to router
