+++ crates/bvs-library/src/addr.rs
@@ -1,5 +1,8 @@
 use cosmwasm_std::{Addr, Api, StdResult};
 
+pub type Service = Addr;
+pub type Operator = Addr;
+
 /// Validate a list of addresses
 /// Return a vector of validated addresses
 pub fn validate_addrs(api: &dyn Api, addr: &[String]) -> StdResult<Vec<Addr>> {
