+++ crates/bvs-registry/Cargo.toml
@@ -26,7 +26,6 @@ cw2 = { workspace = true }
 thiserror = { workspace = true }
 serde = { workspace = true }
 schemars = { workspace = true }
-sha3 = { workspace = true }
 
 bvs-library = { workspace = true }
 bvs-pauser = { workspace = true }
