+++ crates/bvs-vault-router/Cargo.toml
@@ -30,6 +30,7 @@ schemars = { workspace = true }
 bvs-library = { workspace = true }
 bvs-registry = { workspace = true }
 bvs-pauser = { workspace = true }
+sha3 = { workspace = true }
 
 [target."cfg(not(target_arch = \"wasm32\"))".dependencies]
 cw-multi-test = { workspace = true }
